Only show apps on dashboard if they have a home page
Trello: https://trello.com/c/VLXpwPJ8

Previously the Asset Manager & Content Preview apps were not displayed
on the dashboard, because `Doorkeeper::Application#show_on_dashboard?`
was set to `false`. When we converted that attribute into
`Doorkeeper::Application#api_only?` in this commit [1] as part of this
PR [2], those two apps has `#api_only?` set to `true` which prevented
them from being displayed on the dashboard.

However, other changes in [2] meant that permission checkboxes for those
apps were not displayed on any permission-related forms, because they
were marked as API-only. Since the Rails controllers make use of the
`supported_permission_ids` array param, this meant that when a user's
permissions were updated, the permissions for Asset Manager & Content
Preview were inadvertently removed.

Also for similar reasons, since the changes in this PR [3], signin
permissions for Asset Manager & Content Preview apps (which are included
in the list of default permissions) were not being added to new users.

It turns out that signin permission for Asset Manager & Content Preview
are needed for users to access draft content & draft assets, i.e. while
they don't have a home page, these apps do serve content to non-API
users and hence while they do have an API, they are not API-only apps.
This was a real problem for new users never getting access to draft
content/assets and for updated users having access to draft
content/assets removed.

To fix this problem, we marked Asset Manager & Content Preview as not
API-only, so that their permissions do now show up on permission-related
forms. However, this in turn meant the two apps were displayed on the
dashboard, even though they don't have a home page.

I contemplated re-introducing
`Doorkeeper::Application#show_on_dashboard?`, but in the end I decided
to make use of the existing `Doorkeeper::Application#home_uri` attribute
which does not have to be set and only seems to be used for displaying
the link to the app on the dashboard (not for any OAuth-related
malarky). So now any application which does not have `#home_uri` set is
not displayed on the dashboard. I've expanded the "hint" text for
`Doorkeeper::Application#home_uri` to explain this. I plan to remove the
`#home_uri` values for Asset Manager & Content Preview in integration &
production via the UI once these changes have been merged.

Note that I've had to make the `Doorkeeper::Application#substituted_uri`
method more complicated by adding a guard condition which isn't very
nice. However, I believe that functionality relates to the migration to
the EKS platform and can be removed, so I'm not overly concerned about
complicating that method.

[1]: b21bee195e8a1663dae0014d80f74c49d121a188
[2]: https://github.com/alphagov/signon/pull/2452
[3]: https://github.com/alphagov/signon/pull/2451