Remove redundant EKS redirect for application URIs
This effectively reverts the changes in #1911 which seem to relate to
the migration of applications to the EKS platform. This migration
happened about a year ago and as far as I can see the relevant env vars
are not currently set:

* https://github.com/search?q=repo%3Aalphagov%2Fgovuk-helm-charts%20SIGNON_APPS_URI_SUB_PATTERN&type=code
* https://github.com/search?q=repo%3Aalphagov%2Fgovuk-helm-charts+SIGNON_APPS_URI_SUB_REPLACEMENT&type=code

In fact, it's not obvious the two env vars have *ever* been set:

* https://github.com/search?q=org%3Aalphagov+SIGNON_APPS_URI_SUB_PATTERN&type=code
* https://github.com/search?q=org%3Aalphagov+SIGNON_APPS_URI_SUB_REPLACEMENT&type=code

Either way, since they're not currently set, this functionality is
redudant and unnecessarily complicating the code.

I've also removed the two env vars from the docs which were added more
recently.