Rename Account::PasswordsController#update_password -> #update
Now that we've split the forms for updating a user's email and password
into separate controllers, there's no need for the `_email` suffix in
the name of this action. Furthermore, this makes the controller more
like a standard Rails RESTful controller which simplifies the routes a
bit and makes the code a bit less surprising.