Rename Account::PasswordsController#show -> #edit
I found it confusing that this action was named `show` when it's main
purpose is to render a form which is used to send a request to the
`update` action. It's more idiomatic and less surprising to name the
action `edit`.

Note that this does change the URL path from `/account/password` ->
`/account/password/edit`, but I haven't seen any evidence that we care
what the URL is.