Rename Account::OrganisationsController#show -> #edit
I found it confusing that this action was named `show` when it's main
purpose is to render a form which is used to send a request to the
`update` action. It's more idiomatic and less surprising to name the
action `edit`.

It might make sense to split the read-only version of this page out into
a separate `show` action & template, but I'm going to leave that for
now.

Note that this does change the URL path from `/account/organisation` ->
`/account/organisation/edit`, but I haven't seen any evidence that we
care what the URL is.