Display error on email input on edit acccount email page
Previously when a validation error occurred we were displaying an error
summary component, but no error message components against the
individual input field as recommended [1] in the Design System.

Supplying `error_items` to the `input` component fixes this and
highlights the input field in red when there is an error.

[1]: https://design-system.service.gov.uk/patterns/validation/#how-to-tell-the-user-about-validation-errors