Display error on inputs on edit password pages
Previously when a validation error occurred we were displaying an error
summary component, but no error message components against the
individual input fields as recommended [1] in the Design System.

Supplying `error_items` to the `input` components fixes this and
highlights the input field in red when there is an error.

Note that the `PasswordStrengthIndicator` JS module also adds error
messages dynamically to the `password` & `password_confirmation` inputs.
However, it adds them in a different place (above rather than below the
hint), it doesn't use the `error_message` component [2] for these error
messages, and it doesn't take account of whether the same error message
already exists (e.g. from a previous form submission). This can make the
display of error messages a bit confusing. However, I think this commit
is still an improvement (especially for browsers without JS enabled) and
the right solution is to fix the `PasswordStrengthIndicator` JS module.
I plan to add a Trello card to capture this, because I don't think it's
super urgent.

[1]: https://design-system.service.gov.uk/patterns/validation/#how-to-tell-the-user-about-validation-errors