Display links to fields with errors on edit password pages
Previously, while we were displaying a summary list of the errors at the
top of the page and individual error messages against each of the input
fields, we were not linking the errors in the summary list to the
relevant input field as recommended [1] in the Design System.

Using the `error_summary` component with a appropriate `items` option
rather than the custom markup fixes this problem and makes these pages
more consistent with other pages in the app.

[1]: https://design-system.service.gov.uk/patterns/validation/#how-to-tell-the-user-about-validation-errors