Inline change_password_errors partial
Now that we're using the `error_summary` component, there's less
duplication to worrry about and I want to change the title in one of
these. We already have many other very similar usages of the
`error_summary` component across the codebase, so I think it would be
better to extract the duplication in a more generic way.