Simplify UsersController#load_and_authorize_user
Setting the `user` instance variable to `current_user` if the current
user has the "normal" role was introduced in this commit [1].

It's quite hard to see what was going on in that commit, but it looks as
if prior to the commit there had been a separate
`Admin::UsersController` which was how admin users managed other users;
whereas `UsersController` was specifically for managing your own user.

In the commit, the two controller were effectively combined and so
`UsersController#load_user` (which later became
`#load_and_authorize_user`) had to cope with both scenarios; hence the
assignment with the ternary expression:

    @user = current_user.normal? ? current_user : User.find(params[:id])

This had the effect of only allowing a normal user to edit their own
user, although it must've been a bit weird, because it didn't redirect
and so the URL would've still had the ID of another user even though you
would have been editing your own user.

More recently, in this commit [2], we added an account page and removed
the last use of the `edit_user_path` route helper method in
combination with the `current_user`. This meant that it was no longer
possible to navigate to the edit user page if you were a normal user.
Although obviously still possible if you hacked the URL.

And then, in this commit [3], we started redirecting users who were
trying to edit their own user to the account page which has various
sub-pages which allow a user to edit their own user.

The combination of the assignment with the ternary expression (which we
are removing in this commit) and the redirect added in [3] meant that if
a normal user tried to access the edit user page at all (either for
their own user or another user), then they were redirected to the
account page.

However, the means by which this is achieved is quite confusing. By
removing the assignment with the ternary expression we're changing the
behaviour for normal users very slightly, but hopefully making the code
easier to understand.

The change in behaviour only affects normal users who are trying
(presumably by hacking the URL) to edit another user. Prior to this
commit, they would've been redirected to the account page.  After this
commit, they are redirected to their dashboard and they see a "You do
not have permission to perform this action" flash alert. While this is a
change in behaviour, it doesn't seem unreasonable to me and I think it
makes the code easier to understand.

[1]: https://github.com/alphagov/signon/commit/b730700d82e5de79bc0b3004d74c7e547e7eb362?w=1
[2]: https://github.com/alphagov/signon/commit/18ea94ee84928ba86e56c9f09adb6df6601230dd
[3]: https://github.com/alphagov/signon/commit/acad605426e87bd6c8cdd4f18da7f3da425ce57f