Remove navigation concern from UserPolicy#edit?
This guard condition for when the current user is trying to edit their
own user was added in this commit [1].

Reading the commit notes, my understanding is that there isn't anything
inherently *wrong* with a user editing their own user, it's just that we
want there just to be one canonical place where a user edits their own
user. This seems like a navigational concern rather than a
authorization concern and so I don't think it belongs in the
`UserPolicy` class.

Furthermore, in this prior commit [2], we started redirecting users
trying to edit their own user to their account page and so I feel as if
this navigational concern is already covered.

Since the redirect only applies to `UsersController#edit`, but the logic
in `UserPolicy#edit?` is re-used for many other actions, this change
will mean that if someone can craft e.g. a suitable PUT request to
`UsersController#update` they will *now* be able to successfully update
their own user; whereas previously they would have been kicked back to
the dashboard page with a flash alert saying "You do not have permission
to perform this action". However, this is in line with my earlier
assertion that there isn't anything inherently wrong with a user editing
their own user; it's just that ideally we don't want them to do it
through this controller.

[1]: https://github.com/alphagov/signon/commit/bcde44227c48a7a29cba0d925d6a740587adeb00
[2]: https://github.com/alphagov/signon/commit/acad605426e87bd6c8cdd4f18da7f3da425ce57f