Rename Account::PermissionsController#{index, show}
I want to introduce an "edit" route here which will allow users to
edit a collection of permissions for a given app in a single page. I
think it therefore makes more sense to model this route as a singular
resource acting on the collection of permissions. So instead of

    /account/applications/:id/permissions
    /account/applications/:id/permissions/:id/edit

we want

    /account/applications/:id/permissions
    /account/applications/:id/permissions/edit

This commit renames the `index` action to the `show` action,
allowing me to naturally add the `edit` route in a subsequent commit.