Introduce Account::PermissionsController#edit page
This page will eventually allow the user to edit their permissions for
a given application. We currently only want to allow admins and
publishing managers to be able to edit their own permissions, so
making the `edit_permissions?` policy an alias for
`remove_signin_permission?` achieves this (and also covers the case
where the signin permission for a given app is delegatable).

For now I've put the controller action and policy in place and added
the breadcrumb trail and title to the view. I'll add additional
functionality in subsequent commits.