Allow permissions to be updated using checkboxes
This commit adds checkboxes for each application permission to
`Account::PermissionsController#edit`.

The `update` controller method takes an array of permission names and
passes them to a new method
`User#replace_application_permissions`. I've added this method in
order to cope with checkboxes for existing permissions being unchecked
so that they are removed from the user. Rather than pick through the
permissions one-by-one and action them, I've decided to delete all
permissions and replace them with new ones inside a Transaction.

A consequence of this is that the signin permission can also be
removed for a user via this form. Currently we have a "revoke access"
button on the permissions show page which allows the signin permission
to be modified - so we don't need to include the signin
permission. I'm going to address that in a subsequent commit.

Currently I'm redirecting to the account_applications_path but I
haven't yet set any kind of "success" flash message. I'm going to
handle that in a subsequent commit.

Given that the permissions are pre-defined by the application and this
controller/page simply toggles them, I don't think there's an error
state we need to account for (at least not one that could be caused by
using the UI), so I haven't added any error states.