Use contexts consistently for roles in UsersControllerTest
This test is over 800 lines long and I was struggling to find my way
around it. I think that consistently using 2nd-level contexts for
signing in different current user roles is less suprising.

The diff is best viewed using the `--ignore-all-space` git diff option,
because lots of the indentation has changed.