Make sure signin permission isn't removed when updating
Before this commit if a user updated their permissions and submitted
the form their signin permission was lost (since
User#replace_application_permissions) was replacing all permissions
with the new ones selected in the form.

I've added a hidden field to the form such that he signin permission
is always included in the parameters.