Extract edit user name page
Trello: https://trello.com/c/uZD2I9dj

This is a step on the journey to move the edit user page to use the
GOV.UK Design System. The new design calls for separate pages for
editing the different user attribute and this is the first of them.

The new `app/views/users/names/edit.html.erb` template in this commit is
closely based on the relevant bits of `app/views/users/edit.html.erb` &
`app/views/users/_form_fields.html.erb` and I haven't yet changed it to
use the Design System. I plan to do that in a subsequent commit.

The new `Users::NamesController` is closely based on the relevant bits
of code from `UsersController`, even though some of it is probably
overkill in the new context, e.g. the use of `UserUpdate` &
`UserParameterSanitiser`. However, I thought it was worth keeping this
step as small as possible.

I'm reusing `UserPolicy#edit?` & `UserPolicy#update?` for the
authorization in the new `Users::NamesController`, because that still
seems to make sense.

I've tried to add tests in `Users::NamesControllerTest` to capture the
behaviour implied by how things worked when they were in
`UsersController`. It's not obvious that all of this was previously
captured in `UsersControllerTest`.

Rather than creating a combinatorial explosion of tests in
`Users::NamesControllerTest` relating to whether a user with all the
different roles can edit another user with all the different roles, I've
resorted to stubbing `UserPolicy.new` and relevant predicate methods on
the `UserPolicy` instance. Although this is a bit ugly, since
`UserPolicy` is thoroughly tested in `UserPolicyTest`, it seems like a
pragmatic option.

I've added a single integration test just to cover an obvious "happy
path" usage of the functionality.