Move edit user name page to GOV.UK Design System
Trello: https://trello.com/c/uZD2I9dj

This changes the edit user name page to use the Design System and makes
a few other tweaks to bring it into line with the latest designs.

I've introduced a new `title_caption` `content_for` block in the
`app/views/layouts/admin_layout.html.erb` layout template to support the
new "Manage other users" caption which appears in this type of page in
the latest designs. It feels as if this area of the layout is getting a
bit complicated and it might be better to push a chunk of the relevant
markup down into the individual templates even at the cost of some
duplication. The latter could probably be better avoided by other means,
e.g. helper methods, presenter objects, etc.

The page title & heading are slightly different to match other similar
pages.

I've added breadcrumbs to make this page consistent with other pages
even though they're not shown in the latest design for this type of
page.

I'm using the `error_summary` component in combination with identical
code we've used elsewhere so that the errors in the summary link to the
relevant form field. Similarly I'm using the `input` component with
`error_items` set appropriately, so that any relevant errors are
displayed alongside the `name` field and the field itself is highlighted
in red as per this Design System guidance [1]. I've enhanced a test in
`Users::NamesControllerTest` to cover this new behaviour.

The button text has changed from "Update User" -> "Change name" to match
other similar pages.

[1]: https://design-system.service.gov.uk/patterns/validation/#how-to-tell-the-user-about-validation-errors