Simplify Users::NamesController#user_params
I found the `UserParameterSanitiser` and associated logic that we
brought from the `UsersController` quite confusing somewhat overkill for
this controller, especially as the `.permitted_user_params` methods on
the `Roles::Base` subclasses all return an array including `:name` which
is the only parameter we're interested in in this controller.

I did consider simplifying `Users::NamesController#user_params` furhter,
i.e. to just the idiomatic `params.require(:user).permit(:name)` which
is actually all we need. However, since it's possible one or more of the
`.permitted_user_params` methods on the `Roles::Base` subclasses might
change and stop including `:name`, I thought I ought to handle that
eventuality.

I think ideally we'd make use of Pundit's strong parameters
functionality [1] and move this logic out of the `Roles::Base`
subclasses and `UserParameterSanitiser` into the relevant policy class.
However, that's a job for another day!

[1]: https://github.com/varvet/pundit/blob/4d8cdf1c10058c12f2c175f30b408f5d5532a00b/README.md#strong-parameters