Simplify Users::EmailsController#user_params
c.f. this commit [1] where we did the same for `Users::NamesController`.

I found the `UserParameterSanitiser` and associated logic that we
brought from the `UsersController` quite confusing somewhat overkill for
this controller, especially as the `.permitted_user_params` methods on
the `Roles::Base` subclasses all return an array including `:email` which
is the only parameter we're interested in in this controller.

I did consider simplifying `Users::EmailsController#user_params`
further, i.e. to just the idiomatic
`params.require(:user).permit(:email)` which is actually all we need.
However, since it's possible one or more of the `.permitted_user_params`
methods on the `Roles::Base` subclasses might change and stop including
`:email`, I thought I ought to handle that eventuality.

I think ideally we'd make use of Pundit's strong parameters
functionality [2] and move this logic out of the `Roles::Base`
subclasses and `UserParameterSanitiser` into the relevant policy class.
However, that's a job for another day!

[1]: https://github.com/alphagov/signon/commit/2c8010c059768774257c61cf14a04f515a9a0a14
[2]: https://github.com/varvet/pundit/blob/4d8cdf1c10058c12f2c175f30b408f5d5532a00b/README.md#strong-parameters