Inline UserUpdate#call into Users::NamesController
Only including the few bits we need.

We were hardly using any of the logic in `UserUpdate#call` and so I
think it's simpler and more idiomatic to include the logic directly in
the controller action.

In general I think there are better ways to avoid the kind of
duplication that `UserUpdate` was intended to avoid.