Use original user name in page title, etc on edit user name page
Trello: https://trello.com/c/uwgUF9hF

This fixes a bug that was originally present on the edit user page, but
made its way into the recently added edit user name page. The bug is no
longer a problem on the user edit page, because you can no longer change
the user's name on that page.

The problem is best described by the steps to reproduce it when it was
on the user edit page:

* As an admin of some sort edit a user by clicking on a link from the
  table on the users index page
* See that the page title & heading is something like: "Edit “James
  Mead”"
* Change the name field to be blank
* Click the "Update user" button at the bottom of the page
* See that the page title and heading changes to "Edit “”", i.e. it uses
  the new name value and not the old one.