Change default name in User factory to be less confusing
I've seen the string "A name is now required" on a number of occasions
when trying to debug a failing test and assumed that it related to some
kind of unexpected validation error.

This changes the name to use a sequence a bit like the email attribute
which should be a lot less confusing in similar circumstances.