Fix redirect after admin updates another user's name
Trello: https://trello.com/c/DjdHEqxG

This was a mistake in #2497.

When the update completes successfully, the admin should be taken back
to the edit user page for the user in question.

Previously they were redirected to the *non-existent* path for a
non-existent `UsersController#show` action which was somehow (🤷)
being handled by `UsersController#edit` as I intended!