Extract edit user email page
Trello: https://trello.com/c/SsVuShZm

c.f. this commit [1] where we did the same for user name.

This is a step on the journey to move the edit user page to use the
GOV.UK Design System. The new design calls for separate pages for
editing the different user attribute and this is the next one.

The new `app/views/users/emails/edit.html.erb` template in this commit
is closely based on the relevant bits of `app/views/users/edit.html.erb`
& `app/views/users/_form_fields.html.erb`. I haven't yet moved the
"Resend confirmation email" or "Cancel change" functionality and I
haven't yet changed the template to use the Design System. I plan to do
that in subsequent commits.

The new `Users::EmailsController` is closely based on the relevant bits
of code from `UsersController`, even though some of it is probably
overkill in the new context, e.g. the use of `UserUpdate` &
`UserParameterSanitiser`. However, I thought it was worth keeping this
step as small as possible.

I'm reusing `UserPolicy#edit?` & `UserPolicy#update?` for the
authorization in the new `Users::EmailsController`, because that still
seems to make sense.

I've tried to add tests in `Users::EmailsControllerTest` to capture the
behaviour implied by how things worked when they were in
`UsersController`. It's not obvious that all of this was previously
captured in `UsersControllerTest` or `UserUpdateTest`.

Rather than creating a combinatorial explosion of tests in
`Users::EmailsControllerTest` relating to whether a user with all the
different roles can edit another user with all the different roles, I've
resorted to stubbing `UserPolicy.new` and relevant predicate methods on
the `UserPolicy` instance. Although this is a bit ugly, since
`UserPolicy` is thoroughly tested in `UserPolicyTest`, it seems like a
pragmatic option.

The are already some integration tests in `EmailChangeTest` so I've just
changed them to make them work with the new page.

[1]: https://github.com/alphagov/signon/commit/899a8a16356ad3b9e3434feaf99f5e5f2bf8a243