Move edit user email page to GOV.UK Design System
Trello: https://trello.com/c/SsVuShZm

c.f. this commit [1] where we did the same for user name.

This changes the edit user email page to use the Design System and makes
a few other tweaks to bring it into line with the latest designs.

The page title & heading are slightly different to match other similar
pages.

I've added breadcrumbs to make this page consistent with other pages
even though they're not shown in the latest design for this type of
page.

The template for this controller is a bit more complicated than the one
for editing a user's name, because it includes some functionality for
seeing a user's unconfirmed email, for resending an email change email,
and for cancelling an email change. I've based the design on the similar
functionality in the `app/views/account/emails/edit.html.erb` template,
although I've reworded the text to make it clear that in this case it's
an admin user making changes to a user rather than a user making changes
to their own user. Also note that this unconfirmed email functionality
is only activated when a user changes their own email (through the
relevant account page) and not when an admin changes another users email
- in that case the change is effectively confirmed immediately and
automatically.

I'm using the `error_summary` component in combination with identical
code we've used elsewhere so that the errors in the summary link to the
relevant form field. Similarly I'm using the `input` component with
`error_items` set appropriately, so that any relevant errors are
displayed alongside the `email` field and the field itself is highlighted
in red as per this Design System guidance [2]. I've enhanced a test in
`Users::EmailsControllerTest` to cover this new behaviour.

The button text has changed from "Update User" -> "Change email" to
match other similar pages.

[1]: https://github.com/alphagov/signon/commit/f71a128b6a6f9cd86bd8c08508a35d8780c128ef
[2]: https://design-system.service.gov.uk/patterns/validation/#how-to-tell-the-user-about-validation-errors