Remove irrelevant test from UsersControllerTest
This is orthogonal to the rest of this PR, but I thought I'd make the
change when I noticed it. This test became redundant in this commit [1]
or one of its associated commits in #2484.

This redirect is no longer relevant to a normal user, because they are
not authorized to view the page at all and so they will be redirected to
the dashboard page with a message saying they don't have permission to
access the page.

There is a separate test under the "for the currently logged in user"
context which checks that an authorized (admin) user is redirected to
the account page.

[1]: https://github.com/alphagov/signon/commit/baf42ed8863a86e30f2055eb0606506af86ed5aa