SPIKE: How we might be able to simplify/improve UserUpdate
Using `ActiveSupport::CurrentAttributes` means that we can have access
to the current user in models; not just in controllers. This in turn
means we ought to be able to move all of the logic in `UserUpdate` which
is recording events in the activity log into model callbacks. These
ought to be a lot more reliable in the sense that we don't need to keep
remembering to use `UserUpdate` everywhere.

/cc @chrisroos