Update applications when user's name changes
Trello: https://trello.com/c/2t9zx0Aq

Prior to #2497 changing a user's name in Signon would queue a job to
push that update out to the applications the user has access to. This
commit reinstates the previous behaviour by using `UserUpdate` in
`Users::NamesController#update`, which in turn uses `PermissionUpdater`
to push updates to the user's apps.

I can't find any documentation that describes the changes that result in push
updates to a user's applications, but as far as I can see, changing a
user's name has resulted in a push update since the relevant
functionality was added to Signon[1] and gds-sso[2] in 2012 so it seems
reasonable that we retain this behaviour.

Note that `UserUpdate#record_update`[3] records an `EventLog` event so
I've removed this from `Users::NamesController#update`.

[1]: https://github.com/alphagov/signon/pull/6/commits/f339848ac0c98188b0fc746935324f59215d26f2
[2]: https://github.com/alphagov/gds-sso/commit/8c0888cda78b2d0323da26ff4c8da362074168d7
[3]: https://github.com/alphagov/signon/blob/16299a28ed46d12a8b0cb61c1cf755166317057c/app/services/user_update.rb#L72