Add test for SSO Push updates to emails_controller_test
As far as I can tell we've been sending push updates (to the applications
a user has access to) when a user's email changes since the functionality
was added to Signon[1] and gds-sso[2] in 2012.

Although we're still sending those push updates from the new
`Users::EmailsController` (added in #2509) we weren't testing it in the
controller test. I'm adding a test (copied from the previous commit) to
avoid regressions in future.

[1]: https://github.com/alphagov/signon/pull/6/commits/f339848ac0c98188b0fc746935324f59215d26f2
[2]: https://github.com/alphagov/gds-sso/commit/8c0888cda78b2d0323da26ff4c8da362074168d7