Extract edit user role page
Trello: https://trello.com/c/shaJpu8F

c.f. this commit [1] where we did the same for user name.

This is a step on the journey to move the edit user page to use the
GOV.UK Design System. The new design calls for separate pages for
editing the different user attribute and this is the next one.

The new `app/views/users/roles/edit.html.erb` template in this commit is
closely based on the relevant bits of `app/views/users/edit.html.erb` &
`app/views/users/_form_fields.html.erb`. I haven't yet changed the
template to use the Design System. I plan to do that in subsequent
commits.

The new `Users::RolesController` is closely based on the relevant bits
of code from `UsersController`, even though some of it is probably
overkill in the new context, e.g. the use of `UserUpdate` &
`UserParameterSanitiser`. However, I thought it was worth keeping this
step as small as possible.

I'm reusing `UserPolicy#edit?` & `UserPolicy#update?` for the
authorization in the new `Users::RolesController`, because that still
seems to make sense. However, note that we also make use of the existing
unconventional `UserPolicy#assign_role?` predicate method which does not
correspond to a controller action. I suspect there's some simplification
we could do here to make this more idiomatic, but I've left that for
now.

I've tried to add tests in `Users::RolesControllerTest` to capture the
behaviour implied by how things worked when they were in
`UsersController`. It's not obvious that all of this was previously
captured in `UsersControllerTest` or `UserUpdateTest`. Also note that,
unlike in `Users::NamesControllerTest` & `Users::EmailsControllerTest`,
most of the tests are in the context of a Superadmin user doing the
editing rather than just an Admin user. This is because only Superadmins
can change a user's role (see `UserPolicy#assign_role?`).

It took me a while to understand that the `User#exempt_from_2sv?`
condition in the template is there in order to avoid the validation
error that's added by `User#user_can_be_exempted_from_2sv` when an
attempt is made to upgrade a normal user exempted from 2SV to any admin
role. I have added tests for this scenario for both
`Users::RolesController#edit` and `Users::RolesController#update`, even
though it's unlikely that a user would ever see the relevant validation
error unless they hacked the form.

Rather than creating a combinatorial explosion of tests in
`Users::RolesControllerTest` relating to whether a user with all the
different roles can edit another user with all the different roles, I've
resorted to stubbing `UserPolicy.new` and relevant predicate methods on
the `UserPolicy` instance. Although this is a bit ugly, since
`UserPolicy` is thoroughly tested in `UserPolicyTest`, it seems like a
pragmatic option.

The are already some integration tests in `ChangeUserRoleTest` so I've
just changed them to make them work with the new page.

[1]: https://github.com/alphagov/signon/commit/899a8a16356ad3b9e3434feaf99f5e5f2bf8a243