Move edit user role page to GOV.UK Design System
Trello: https://trello.com/c/shaJpu8F

c.f. this commit [1] where we did the same for user name.

This changes the edit user role page to use the Design System and makes
a few other tweaks to bring it into line with the latest designs.

The page title & heading are slightly different to match other similar
pages.

I've added breadcrumbs to make this page consistent with other pages
even though they're not shown in the latest design for this type of
page.

The template for this controller is a bit more complicated than the one
for editing a user's name, because it has to handle the case where the
user being edited is exempt from 2SV which means they cannot be upgraded
from the normal role to any other (admin) role. In this case we just
display some inset text to explain why the user's role cannot be
changed. I'm not sure this is a great user experience, but it's
somewhate similar to what happens in the account change role page. Note
that I've humanized the role in the inset text, because I think that's
more consistent with how we're displaying role names to users in the
rest of the app.

I did contemplate hiding the "Change (role)" link on the edit user page
in this scenario. However, this link is already hidden when the current
user does not have permission to change a user's role and I thought it
would be more confusing if there were two reasons why it might not show
up. I think the behaviour in this commit is good enough for now and we
can iterate on the UX/design later.

I've based the design on the similar functionality in the
`app/views/account/roles/edit.html.erb` template, although I've reworded
the text to make it clear that in this case it's an admin user making
changes to a user rather than a user making changes to their own user.

I've also included the same hint HTML as used for a very similar `role`
select component in `app/views/devise/invitations/new.html.erb` which
replaces the "help-block" `span` element. I plan to extract this
duplication in a subsequent commit.

I'm using the `error_summary` component in combination with identical
code we've used elsewhere so that the errors in the summary link to the
relevant form field. Similarly I'm using the `select` component with
`error_message` set appropriately, so that any relevant errors are
displayed alongside the `role` field and the field itself is highlighted
in red as per this Design System guidance [2]. I've enhanced a test in
`Users::RolesControllerTest` to cover this new behaviour. Note that in
normal operation there are unlikely to be validation errors on this
page. However, it is possible if someone hacks the form or e.g. you try
to upgrade a user without an organisation to Organisation Admin.

The button text has changed from "Update User" -> "Change role" and I've
introduced a "Cancel" link to match other similar pages.

[1]: https://github.com/alphagov/signon/commit/f71a128b6a6f9cd86bd8c08508a35d8780c128ef
[2]: https://design-system.service.gov.uk/patterns/validation/#how-to-tell-the-user-about-validation-errors